Preallocate chunk result array upfront
Previously, np.concatenate was used to aggregate all row runs in a chunk.
This meant that we needed twice us much memory - one set for all row
runs and one for the concatenate result. Now, the result is
preallocated, and `pyrap.tables.table.getcolnp` is used to dump data
directly into the result array.